metadata: Remove compressed object field validation
The field validation done after parsing the metadata did not server a useful
propose, it validate that the code we had implanted for compressed objects
had the proper fields for parsing, but did nothing to validate the metadata
received. It also had a bug when parsing compressed Raw type objects. The block
is now fully removed

Signed-off-by: Jonathas-Conceicao <jonathas.conceicao@ossystems.com.br>